+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # CHANGELOG
 
+## 0.2.0.dev (2020-12-07)
+
+### Added
+- Graphical User Interface [#7](https://github.com/nekeal/platonic-io/pull/7)
+- Plate recognition engine [#8](https://github.com/nekeal/platonic-io/pull/8)
+- Additional CLI [#8](https://github.com/nekeal/platonic-io/pull/8)
+
 
 ## 0.1.0 (2020-12-01)
 
 ### Added
 - `LicencePlateOCRReader` class for reading text from license plates [#3](https://github.com/nekeal/platonic-io/pull/3)
 - Command `platonic-io ocr` for `LicencePlateOCRReader` [#3](https://github.com/nekeal/platonic-io/pull/3)
-
-
-## 0.2.0.dev (2020-xx-xx)